+++ exporters/otlp/otlp_metric_test.go
@@ -58,7 +58,7 @@ type checkpointSet struct {
 	records []metricsdk.Record
 }
 
-func (m *checkpointSet) ForEach(_ metricsdk.ExportKindSelector, fn func(metricsdk.Record) error) error {
+func (m *checkpointSet) ForEach(_ metricsdk.AggregationTemporalitySelector, fn func(metricsdk.Record) error) error {
 	for _, r := range m.records {
 		if err := fn(r); err != nil && err != aggregation.ErrNoData {
 			return err
@@ -676,7 +676,7 @@ func TestResourceInstLibMetricGroupingExport(t *testing.T) {
 	)
 }
 
-func TestStatelessExportKind(t *testing.T) {
+func TestStatelessAggregationTemporality(t *testing.T) {
 	type testcase struct {
 		name           string
 		instrumentKind metric.InstrumentKind
@@ -694,8 +694,8 @@ func TestStatelessExportKind(t *testing.T) {
 			runMetricExportTests(
 				t,
 				[]otlp.ExporterOption{
-					otlp.WithMetricExportKindSelector(
-						metricsdk.StatelessExportKindSelector(),
+					otlp.WithMetricAggregationTemporalitySelector(
+						metricsdk.StatelessAggregationTemporalitySelector(),
 					),
 				},
 				[]record{