+++ exporters/stdout/metric.go
@@ -52,8 +52,8 @@ type quantile struct {
 	Value    interface{} `json:"Value"`
 }
 
-func (e *metricExporter) ExportKindFor(desc *metric.Descriptor, kind aggregation.Kind) exportmetric.ExportKind {
-	return exportmetric.StatelessExportKindSelector().ExportKindFor(desc, kind)
+func (e *metricExporter) AggregationTemporalityFor(desc *metric.Descriptor, kind aggregation.Kind) exportmetric.AggregationTemporality {
+	return exportmetric.StatelessAggregationTemporalitySelector().AggregationTemporalityFor(desc, kind)
 }
 
 func (e *metricExporter) Export(_ context.Context, checkpointSet exportmetric.CheckpointSet) error {