+++ sdk/metric/controller/push/push_test.go
@@ -66,7 +66,7 @@ func init() {
 
 func newExporter() *processorTest.Exporter {
 	return processorTest.NewExporter(
-		export.StatelessExportKindSelector(),
+		export.StatelessAggregationTemporalitySelector(),
 		label.DefaultEncoder(),
 	)
 }