+++ sdk/metric/processor/processortest/test_test.go
@@ -74,7 +74,7 @@ func TestProcessorTesting(t *testing.T) {
 
 	// Export the data and validate it again.
 	exporter := processorTest.NewExporter(
-		export.StatelessExportKindSelector(),
+		export.StatelessAggregationTemporalitySelector(),
 		label.DefaultEncoder(),
 	)